refactor(backend): 多模块架构调整——文件服务/平台统计 + 支付基础设施迁移脚本

+++ easyorange-backend/easyorange-framework/src/main/java/com/cartethyia/easyorange/framework/aspectj/OperLogAspect.java
@@ -179,7 +179,7 @@ public class OperLogAspect {
         operLog.setOperTime(LocalDateTime.now());
 
         Long startTime = (Long) request.getAttribute("requestStartTime");
-        operLog.setCostTime(System.currentTimeMillis() - (startTime != null ? startTime : System.currentTimeMillis()));
+        operLog.setCostTime((int)(System.currentTimeMillis() - (startTime != null ? startTime : System.currentTimeMillis())));
 
         return operLog;
     }
